Modularize the Text-to-PDF feature (#888)
