--- app/src/main/java/swati4star/createpdf/fragment/TextToPdfFragment.java
@@ -1,529 +1,0 @@
-package swati4star.createpdf.fragment;
-
-import android.app.Activity;
-import android.content.Context;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Environment;
-import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.text.Editable;
-import android.text.TextWatcher;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.CheckBox;
-import android.widget.EditText;
-import android.widget.RadioButton;
-import android.widget.RadioGroup;
-
-import com.afollestad.materialdialogs.DialogAction;
-import com.afollestad.materialdialogs.MaterialDialog;
-import com.dd.morphingbutton.MorphingButton;
-import com.github.danielnilsson9.colorpickerview.view.ColorPickerView;
-import com.itextpdf.text.Font;
-
-import java.util.ArrayList;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import swati4star.createpdf.R;
-import swati4star.createpdf.adapter.EnhancementOptionsAdapter;
-import swati4star.createpdf.interfaces.OnItemClickListener;
-import swati4star.createpdf.interfaces.OnTextToPdfInterface;
-import swati4star.createpdf.model.EnhancementOptionsEntity;
-import swati4star.createpdf.model.TextToPDFOptions;
-import swati4star.createpdf.util.ColorUtils;
-import swati4star.createpdf.util.Constants;
-import swati4star.createpdf.util.DialogUtils;
-import swati4star.createpdf.util.DirectoryUtils;
-import swati4star.createpdf.util.FileUtils;
-import swati4star.createpdf.util.MorphButtonUtility;
-import swati4star.createpdf.util.PageSizeUtils;
-import swati4star.createpdf.util.PermissionsUtils;
-import swati4star.createpdf.util.StringUtils;
-import swati4star.createpdf.util.TextToPDFUtils;
-import swati4star.createpdf.util.TextToPdfAsync;
-
-import static android.app.Activity.RESULT_OK;
-import static swati4star.createpdf.util.Constants.DEFAULT_PAGE_COLOR;
-import static swati4star.createpdf.util.Constants.READ_WRITE_PERMISSIONS;
-import static swati4star.createpdf.util.TextEnhancementOptionsUtils.getEnhancementOptions;
-
-public class TextToPdfFragment extends Fragment implements OnItemClickListener,
-        OnTextToPdfInterface {
-
-    private Activity mActivity;
-    private FileUtils mFileUtils;
-    private DirectoryUtils mDirectoryUtils;
-
-    private final int mFileSelectCode = 0;
-    private Uri mTextFileUri = null;
-    private String mFontTitle;
-    private int mFontColor;
-    private int mPageColor;
-    private String mFileExtension;
-    private int mFontSize = 0;
-    private int mButtonClicked = 0;
-    private boolean mPasswordProtected = false;
-    private String mPassword;
-    private static final int PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT = 1;
-    private boolean mPermissionGranted = false;
-    private MaterialDialog mMaterialDialog;
-    private String mFileNameWithType = null;
-
-    @BindView(R.id.enhancement_options_recycle_view_text)
-    RecyclerView mTextEnhancementOptionsRecycleView;
-    @BindView(R.id.selectFile)
-    MorphingButton mSelectFile;
-    @BindView(R.id.createtextpdf)
-    MorphingButton mCreateTextPdf;
-
-    private ArrayList<EnhancementOptionsEntity> mTextEnhancementOptionsEntityArrayList;
-    private EnhancementOptionsAdapter mTextEnhancementOptionsAdapter;
-    private SharedPreferences mSharedPreferences;
-    private Font.FontFamily mFontFamily;
-    private MorphButtonUtility mMorphButtonUtility;
-    private String mPath;
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_text_to_pdf, container, false);
-        mPermissionGranted = PermissionsUtils.checkRuntimePermissions(this, READ_WRITE_PERMISSIONS);
-        mSharedPreferences = PreferenceManager.getDefaultSharedPreferences(mActivity);
-        mFontTitle = String.format(getString(R.string.edit_font_size),
-                mSharedPreferences.getInt(Constants.DEFAULT_FONT_SIZE_TEXT, Constants.DEFAULT_FONT_SIZE));
-        mFontFamily = Font.FontFamily.valueOf(mSharedPreferences.getString(Constants.DEFAULT_FONT_FAMILY_TEXT,
-                Constants.DEFAULT_FONT_FAMILY));
-        mFontColor = mSharedPreferences.getInt(Constants.DEFAULT_FONT_COLOR_TEXT,
-                Constants.DEFAULT_FONT_COLOR);
-        mPageColor = mSharedPreferences.getInt(Constants.DEFAULT_PAGE_COLOR_TTP,
-                DEFAULT_PAGE_COLOR);
-        mMorphButtonUtility = new MorphButtonUtility(mActivity);
-        ButterKnife.bind(this, rootView);
-        showEnhancementOptions();
-        mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
-        mCreateTextPdf.setEnabled(false);
-        PageSizeUtils.mPageSize = mSharedPreferences.getString(Constants.DEFAULT_PAGE_SIZE_TEXT,
-                Constants.DEFAULT_PAGE_SIZE);
-        mFontSize = mSharedPreferences.getInt(Constants.DEFAULT_FONT_SIZE_TEXT, Constants.DEFAULT_FONT_SIZE);
-
-        return rootView;
-    }
-
-    /**
-     * Function to show the enhancement options.
-     */
-    private void showEnhancementOptions() {
-        GridLayoutManager mGridLayoutManager = new GridLayoutManager(mActivity, 2);
-        mTextEnhancementOptionsRecycleView.setLayoutManager(mGridLayoutManager);
-        mTextEnhancementOptionsEntityArrayList = getEnhancementOptions(mActivity, mFontTitle, mFontFamily);
-        mTextEnhancementOptionsAdapter = new EnhancementOptionsAdapter(this, mTextEnhancementOptionsEntityArrayList);
-        mTextEnhancementOptionsRecycleView.setAdapter(mTextEnhancementOptionsAdapter);
-    }
-
-    @Override
-    public void onItemClick(int position) {
-        switch (position) {
-            case 0:
-                editFontSize();
-                break;
-            case 1:
-                changeFontFamily();
-                break;
-            case 2:
-                setPageSize();
-                break;
-            case 3:
-                setPassword();
-                break;
-            case 4:
-                setFontColor();
-                break;
-            case 5:
-                setPageColor();
-                break;
-        }
-    }
-
-    private void setPassword() {
-        MaterialDialog.Builder builder = DialogUtils.getInstance().createCustomDialogWithoutContent(mActivity,
-                R.string.set_password);
-        final MaterialDialog dialog = builder
-                .customView(R.layout.custom_dialog, true)
-                .neutralText(R.string.remove_dialog)
-                .build();
-
-        final View positiveAction = dialog.getActionButton(DialogAction.POSITIVE);
-        final View neutralAction = dialog.getActionButton(DialogAction.NEUTRAL);
-        final EditText passwordInput = dialog.getCustomView().findViewById(R.id.password);
-        passwordInput.setText(mPassword);
-        passwordInput.addTextChangedListener(
-                new TextWatcher() {
-                    @Override
-                    public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                    }
-
-                    @Override
-                    public void onTextChanged(CharSequence s, int start, int before, int count) {
-                        positiveAction.setEnabled(s.toString().trim().length() > 0);
-                    }
-
-                    @Override
-                    public void afterTextChanged(Editable input) {
-                    }
-                });
-
-        positiveAction.setOnClickListener(v -> {
-            if (StringUtils.getInstance().isEmpty(passwordInput.getText())) {
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_password_cannot_be_blank);
-            } else {
-                mPassword = passwordInput.getText().toString();
-                mPasswordProtected = true;
-                onPasswordAdded();
-                dialog.dismiss();
-            }
-        });
-
-        if (StringUtils.getInstance().isNotEmpty(mPassword)) {
-            neutralAction.setOnClickListener(v -> {
-                mPassword = null;
-                onPasswordRemoved();
-                mPasswordProtected = false;
-                dialog.dismiss();
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.password_remove);
-            });
-        }
-        dialog.show();
-        positiveAction.setEnabled(false);
-    }
-
-    private void setPageSize() {
-        PageSizeUtils utils = new PageSizeUtils(mActivity);
-        utils.showPageSizeDialog(false);
-    }
-
-    /**
-     * Shows dialog to change font size
-     */
-    private void changeFontFamily() {
-        String fontFamily = mSharedPreferences.getString(Constants.DEFAULT_FONT_FAMILY_TEXT,
-                Constants.DEFAULT_FONT_FAMILY);
-        int ordinal = Font.FontFamily.valueOf(fontFamily).ordinal();
-        MaterialDialog materialDialog = new MaterialDialog.Builder(mActivity)
-                .title(String.format(getString(R.string.default_font_family_text), fontFamily))
-                .customView(R.layout.dialog_font_family, true)
-                .positiveText(R.string.ok)
-                .negativeText(R.string.cancel)
-                .onPositive((dialog, which) -> {
-                    View view = dialog.getCustomView();
-                    RadioGroup radioGroup = view.findViewById(R.id.radio_group_font_family);
-                    int selectedId = radioGroup.getCheckedRadioButtonId();
-                    RadioButton radioButton = view.findViewById(selectedId);
-                    String fontFamily1 = radioButton.getText().toString();
-                    mFontFamily = Font.FontFamily.valueOf(fontFamily1);
-                    final CheckBox cbSetDefault = view.findViewById(R.id.cbSetDefault);
-                    if (cbSetDefault.isChecked()) {
-                        SharedPreferences.Editor editor = mSharedPreferences.edit();
-                        editor.putString(Constants.DEFAULT_FONT_FAMILY_TEXT, fontFamily1);
-                        editor.apply();
-                    }
-                    showFontFamily();
-                })
-                .build();
-        RadioGroup radioGroup = materialDialog.getCustomView().findViewById(R.id.radio_group_font_family);
-        RadioButton rb = (RadioButton) radioGroup.getChildAt(ordinal);
-        rb.setChecked(true);
-        materialDialog.show();
-    }
-
-    private void setFontColor() {
-        MaterialDialog materialDialog = new MaterialDialog.Builder(mActivity)
-                .title(R.string.font_color)
-                .customView(R.layout.dialog_color_chooser, true)
-                .positiveText(R.string.ok)
-                .negativeText(R.string.cancel)
-                .onPositive((dialog, which) -> {
-                    View view = dialog.getCustomView();
-                    ColorPickerView colorPickerView = view.findViewById(R.id.color_picker);
-                    CheckBox defaultCheckbox = view.findViewById(R.id.set_default);
-                    mFontColor = colorPickerView.getColor();
-                    if (ColorUtils.getInstance().colorSimilarCheck(mFontColor, mPageColor)) {
-                        StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_color_too_close);
-                    }
-                    if (defaultCheckbox.isChecked()) {
-                        SharedPreferences.Editor editor = mSharedPreferences.edit();
-                        editor.putInt(Constants.DEFAULT_FONT_COLOR_TEXT, mFontColor);
-                        editor.apply();
-                    }
-                })
-                .build();
-        ColorPickerView colorPickerView = materialDialog.getCustomView().findViewById(R.id.color_picker);
-        colorPickerView.setColor(mFontColor);
-        materialDialog.show();
-    }
-
-    private void setPageColor() {
-        MaterialDialog materialDialog = new MaterialDialog.Builder(mActivity)
-                .title(R.string.page_color)
-                .customView(R.layout.dialog_color_chooser, true)
-                .positiveText(R.string.ok)
-                .negativeText(R.string.cancel)
-                .onPositive((dialog, which) -> {
-                    View view = dialog.getCustomView();
-                    ColorPickerView colorPickerView = view.findViewById(R.id.color_picker);
-                    CheckBox defaultCheckbox = view.findViewById(R.id.set_default);
-                    mPageColor = colorPickerView.getColor();
-                    if (ColorUtils.getInstance().colorSimilarCheck(mFontColor, mPageColor)) {
-                        StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_color_too_close);
-                    }
-                    if (defaultCheckbox.isChecked()) {
-                        SharedPreferences.Editor editor = mSharedPreferences.edit();
-                        editor.putInt(Constants.DEFAULT_PAGE_COLOR_TTP, mPageColor);
-                        editor.apply();
-                    }
-                })
-                .build();
-        ColorPickerView colorPickerView = materialDialog.getCustomView().findViewById(R.id.color_picker);
-        colorPickerView.setColor(mPageColor);
-        materialDialog.show();
-    }
-
-    /**
-     * Function to take the font size of pdf as user input
-     */
-    private void editFontSize() {
-        new MaterialDialog.Builder(mActivity)
-                .title(mFontTitle)
-                .customView(R.layout.dialog_font_size, true)
-                .positiveText(R.string.ok)
-                .negativeText(R.string.cancel)
-                .onPositive((dialog, which) -> {
-                    final EditText fontInput = dialog.getCustomView().findViewById(R.id.fontInput);
-                    final CheckBox cbSetDefault = dialog.getCustomView().findViewById(R.id.cbSetFontDefault);
-                    try {
-                        int check = Integer.parseInt(String.valueOf(fontInput.getText()));
-                        if (check > 1000 || check < 0) {
-                            StringUtils.getInstance().showSnackbar(mActivity, R.string.invalid_entry);
-                        } else {
-                            mFontSize = check;
-                            showFontSize();
-                            StringUtils.getInstance().showSnackbar(mActivity, R.string.font_size_changed);
-                            if (cbSetDefault.isChecked()) {
-                                SharedPreferences.Editor editor = mSharedPreferences.edit();
-                                editor.putInt(Constants.DEFAULT_FONT_SIZE_TEXT, mFontSize);
-                                editor.apply();
-                                mFontTitle = String.format(getString(R.string.edit_font_size),
-                                        mSharedPreferences.getInt(Constants.DEFAULT_FONT_SIZE_TEXT,
-                                                Constants.DEFAULT_FONT_SIZE));
-                            }
-                        }
-                    } catch (NumberFormatException e) {
-                        StringUtils.getInstance().showSnackbar(mActivity, R.string.invalid_entry);
-                    }
-                })
-                .show();
-    }
-
-    /**
-     * Displays font family in UI
-     */
-    private void showFontFamily() {
-        mTextEnhancementOptionsEntityArrayList.get(1)
-                .setName(getString(R.string.font_family_text) + mFontFamily.name());
-        mTextEnhancementOptionsAdapter.notifyDataSetChanged();
-    }
-
-    /**
-     * Displays font size in UI
-     */
-    private void showFontSize() {
-        mTextEnhancementOptionsEntityArrayList.get(0)
-                .setName(String.format(getString(R.string.font_size), String.valueOf(mFontSize)));
-        mTextEnhancementOptionsAdapter.notifyDataSetChanged();
-    }
-
-    @OnClick(R.id.createtextpdf)
-    public void openCreateTextPdf() {
-        new MaterialDialog.Builder(mActivity)
-                .title(R.string.creating_pdf)
-                .content(R.string.enter_file_name)
-                .input(getString(R.string.example), mFileNameWithType,
-                        (dialog, input) -> {
-                            if (StringUtils.getInstance().isEmpty(input)) {
-                                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_name_not_blank);
-                            } else {
-                                final String inputName = input.toString();
-                                if (!mFileUtils.isFileExist(inputName + getString(R.string.pdf_ext))) {
-                                    createPdf(inputName);
-                                } else {
-                                    MaterialDialog.Builder builder = DialogUtils.getInstance()
-                                            .createOverwriteDialog(mActivity);
-                                    builder.onPositive((dialog12, which) -> createPdf(inputName))
-                                            .onNegative((dialog1, which) -> openCreateTextPdf())
-                                            .show();
-                                }
-                            }
-                        })
-                .show();
-    }
-
-    /**
-     * function to create PDF
-     *
-     * @param mFilename name of file to be created.
-     */
-    private void createPdf(String mFilename) {
-        mPath = mDirectoryUtils.getOrCreatePdfDirectory().getPath();
-        mPath = mPath + "/" + mFilename + mActivity.getString(R.string.pdf_ext);
-        TextToPDFOptions options = new TextToPDFOptions(mFilename, PageSizeUtils.mPageSize, mPasswordProtected,
-                mPassword, mTextFileUri, mFontSize, mFontFamily, mFontColor, mPageColor);
-        TextToPDFUtils fileUtil = new TextToPDFUtils(mActivity);
-        new TextToPdfAsync(fileUtil, options, mFileExtension,
-                TextToPdfFragment.this).execute();
-    }
-
-    /**
-     * Create a file picker to get text file.
-     */
-    @OnClick(R.id.selectFile)
-    public void selectTextFile() {
-        if (!mPermissionGranted) {
-            getRuntimePermissions();
-            return;
-        }
-        if (mButtonClicked == 0) {
-            Uri uri = Uri.parse(Environment.getRootDirectory() + "/");
-            Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
-            intent.setDataAndType(uri, "*/*");
-            String[] mimeTypes = {"application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-                    "application/msword", getString(R.string.text_type)};
-            intent.putExtra(Intent.EXTRA_MIME_TYPES, mimeTypes);
-            intent.addCategory(Intent.CATEGORY_OPENABLE);
-            try {
-                startActivityForResult(
-                        Intent.createChooser(intent, String.valueOf(R.string.select_file)),
-                        mFileSelectCode);
-            } catch (android.content.ActivityNotFoundException ex) {
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.install_file_manager);
-            }
-            mButtonClicked = 1;
-        }
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        mButtonClicked = 0;
-        if (requestCode == mFileSelectCode) {
-            if (resultCode == RESULT_OK) {
-                mTextFileUri = data.getData();
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.text_file_selected);
-                String fileName = mFileUtils.getFileName(mTextFileUri);
-                if (fileName != null) {
-                    if (fileName.endsWith(Constants.textExtension))
-                        mFileExtension = Constants.textExtension;
-                    else if (fileName.endsWith(Constants.docxExtension))
-                        mFileExtension = Constants.docxExtension;
-                    else if (fileName.endsWith(Constants.docExtension))
-                        mFileExtension = Constants.docExtension;
-                    else {
-                        StringUtils.getInstance().showSnackbar(mActivity, R.string.extension_not_supported);
-                        return;
-                    }
-                }
-                mFileNameWithType = mFileUtils.stripExtension(fileName) + getString(R.string.pdf_suffix);
-                mSelectFile.setText(getString(R.string.text_file_name) + fileName);
-                mCreateTextPdf.setEnabled(true);
-                mMorphButtonUtility.morphToSquare(mCreateTextPdf, mMorphButtonUtility.integer());
-            }
-        }
-        super.onActivityResult(requestCode, resultCode, data);
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof Activity) {
-            mActivity = (Activity) context;
-        } else {
-            mActivity = getActivity();
-        }
-
-        mFileUtils = new FileUtils(mActivity);
-        mDirectoryUtils = new DirectoryUtils(mActivity);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           @NonNull String[] permissions, @NonNull int[] grantResults) {
-        if (grantResults.length < 1)
-            return;
-        if (requestCode == PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT) {
-            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                mPermissionGranted = true;
-                selectTextFile();
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_permissions_given);
-            } else
-                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_insufficient_permissions);
-        }
-    }
-
-    private void onPasswordAdded() {
-        mTextEnhancementOptionsEntityArrayList.get(3)
-                .setImage(mActivity.getResources().getDrawable(R.drawable.baseline_done_24));
-        mTextEnhancementOptionsAdapter.notifyDataSetChanged();
-    }
-
-    private void onPasswordRemoved() {
-        mTextEnhancementOptionsEntityArrayList.get(3)
-                .setImage(mActivity.getResources().getDrawable(R.drawable.baseline_enhanced_encryption_24));
-        mTextEnhancementOptionsAdapter.notifyDataSetChanged();
-    }
-
-    private void getRuntimePermissions() {
-        PermissionsUtils.requestRuntimePermissions(
-                mActivity,
-                READ_WRITE_PERMISSIONS,
-                PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT
-        );
-    }
-
-    @Override
-    public void onPDFCreationStarted() {
-        mMaterialDialog = DialogUtils.getInstance().createAnimationDialog(mActivity);
-        mMaterialDialog.show();
-    }
-
-    @Override
-    public void onPDFCreated(boolean success) {
-        if (mMaterialDialog != null && mMaterialDialog.isShowing())
-            mMaterialDialog.dismiss();
-        if (!success) {
-            StringUtils.getInstance().showSnackbar(mActivity, R.string.error_occurred);
-            mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
-            mCreateTextPdf.setEnabled(false);
-            mTextFileUri = null;
-            mButtonClicked = 0;
-            return;
-        }
-        StringUtils.getInstance().getSnackbarwithAction(mActivity, R.string.snackbar_pdfCreated)
-                .setAction(R.string.snackbar_viewAction, v -> mFileUtils.openFile(mPath)).show();
-        mSelectFile.setText(R.string.select_text_file);
-        mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
-        mCreateTextPdf.setEnabled(false);
-        mTextFileUri = null;
-        mButtonClicked = 0;
-        mFontColor = mSharedPreferences.getInt(Constants.DEFAULT_FONT_COLOR_TEXT,
-                Constants.DEFAULT_FONT_COLOR);
-        mPageColor = mSharedPreferences.getInt(Constants.DEFAULT_PAGE_COLOR_TTP,
-                DEFAULT_PAGE_COLOR);
-    }
-}
