+++ app/src/main/java/swati4star/createpdf/fragment/texttopdf/PageSizeEnhancer.java
@@ -1,0 +1,39 @@
+package swati4star.createpdf.fragment.texttopdf;
+
+import android.content.Context;
+import android.preference.PreferenceManager;
+import android.support.annotation.NonNull;
+
+import swati4star.createpdf.R;
+import swati4star.createpdf.interfaces.Enhancer;
+import swati4star.createpdf.model.EnhancementOptionsEntity;
+import swati4star.createpdf.util.Constants;
+import swati4star.createpdf.util.PageSizeUtils;
+
+/**
+ * An {@link Enhancer} that lets you select page size.
+ */
+public class PageSizeEnhancer implements Enhancer {
+
+    private final PageSizeUtils mPageSizeUtils;
+    private final EnhancementOptionsEntity mEnhancementOptionsEntity;
+
+    PageSizeEnhancer(@NonNull final Context context) {
+        mPageSizeUtils = new PageSizeUtils(context);
+        mEnhancementOptionsEntity = new EnhancementOptionsEntity(
+                context, R.drawable.ic_page_size_24dp, R.string.set_page_size_text);
+        PageSizeUtils.mPageSize = PreferenceManager.getDefaultSharedPreferences(context)
+                .getString(Constants.DEFAULT_PAGE_SIZE_TEXT, Constants.DEFAULT_PAGE_SIZE);
+    }
+
+    @Override
+    public void enhance() {
+        mPageSizeUtils.showPageSizeDialog(false);
+    }
+
+    @Override
+    public EnhancementOptionsEntity getEnhancementOptionsEntity() {
+        return mEnhancementOptionsEntity;
+    }
+
+}
