+++ app/src/main/java/swati4star/createpdf/fragment/texttopdf/TextToPdfContract.java
@@ -1,0 +1,17 @@
+package swati4star.createpdf.fragment.texttopdf;
+
+/**
+ * The {@link TextToPdfContract} is a contract used by the fragment to communicate with its
+ * enhancements.
+ */
+public interface TextToPdfContract {
+    /**
+     * Represents the view (the fragment in this case).
+     */
+    interface View {
+        /**
+         * Update the view when enhancement is changed.
+         */
+        void updateView();
+    }
+}
