+++ app/src/main/java/swati4star/createpdf/fragment/texttopdf/TextToPdfFragment.java
@@ -1,0 +1,296 @@
+package swati4star.createpdf.fragment.texttopdf;
+
+import android.app.Activity;
+import android.content.Context;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
+import android.os.Bundle;
+import android.os.Environment;
+import android.support.annotation.NonNull;
+import android.support.v4.app.Fragment;
+import android.support.v7.widget.GridLayoutManager;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.afollestad.materialdialogs.MaterialDialog;
+import com.dd.morphingbutton.MorphingButton;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+import butterknife.OnClick;
+import swati4star.createpdf.R;
+import swati4star.createpdf.adapter.EnhancementOptionsAdapter;
+import swati4star.createpdf.interfaces.Enhancer;
+import swati4star.createpdf.interfaces.OnItemClickListener;
+import swati4star.createpdf.interfaces.OnTextToPdfInterface;
+import swati4star.createpdf.model.EnhancementOptionsEntity;
+import swati4star.createpdf.model.TextToPDFOptions;
+import swati4star.createpdf.util.Constants;
+import swati4star.createpdf.util.DialogUtils;
+import swati4star.createpdf.util.DirectoryUtils;
+import swati4star.createpdf.util.FileUtils;
+import swati4star.createpdf.util.MorphButtonUtility;
+import swati4star.createpdf.util.PageSizeUtils;
+import swati4star.createpdf.util.PermissionsUtils;
+import swati4star.createpdf.util.StringUtils;
+import swati4star.createpdf.util.TextToPDFUtils;
+import swati4star.createpdf.util.TextToPdfAsync;
+
+import static android.app.Activity.RESULT_OK;
+import static swati4star.createpdf.util.Constants.READ_WRITE_PERMISSIONS;
+
+public class TextToPdfFragment extends Fragment implements OnItemClickListener,
+        OnTextToPdfInterface, TextToPdfContract.View {
+
+    private Activity mActivity;
+    private FileUtils mFileUtils;
+    private DirectoryUtils mDirectoryUtils;
+
+    private final int mFileSelectCode = 0;
+    private Uri mTextFileUri = null;
+    private String mFileExtension;
+    private int mButtonClicked = 0;
+    private static final int PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT = 1;
+    private boolean mPermissionGranted = false;
+    private MaterialDialog mMaterialDialog;
+    private String mFileNameWithType = null;
+
+    @BindView(R.id.enhancement_options_recycle_view_text)
+    RecyclerView mTextEnhancementOptionsRecycleView;
+    @BindView(R.id.selectFile)
+    MorphingButton mSelectFile;
+    @BindView(R.id.createtextpdf)
+    MorphingButton mCreateTextPdf;
+
+    private EnhancementOptionsAdapter mTextEnhancementOptionsAdapter;
+    private MorphButtonUtility mMorphButtonUtility;
+    private String mPath;
+    private List<Enhancer> mEnhancerList;
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.fragment_text_to_pdf, container, false);
+        mPermissionGranted = PermissionsUtils.checkRuntimePermissions(this, READ_WRITE_PERMISSIONS);
+
+        mMorphButtonUtility = new MorphButtonUtility(mActivity);
+        ButterKnife.bind(this, rootView);
+
+        addEnhancements();
+        showEnhancementOptions();
+        mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
+        mCreateTextPdf.setEnabled(false);
+        return rootView;
+    }
+
+    private void addEnhancements() {
+        mEnhancerList = new ArrayList<>();
+        for (final Enhancers enhancer: Enhancers.values()) {
+            mEnhancerList.add(enhancer.getEnhancer(mActivity, this));
+        }
+    }
+
+    /**
+     * Function to show the enhancement options.
+     */
+    private void showEnhancementOptions() {
+        GridLayoutManager mGridLayoutManager = new GridLayoutManager(mActivity, 2);
+        mTextEnhancementOptionsRecycleView.setLayoutManager(mGridLayoutManager);
+        List<EnhancementOptionsEntity> optionsEntityist = new ArrayList<>();
+        for (Enhancer enhancer: mEnhancerList) {
+            optionsEntityist.add(enhancer.getEnhancementOptionsEntity());
+        }
+        mTextEnhancementOptionsAdapter = new EnhancementOptionsAdapter(this, optionsEntityist);
+        mTextEnhancementOptionsRecycleView.setAdapter(mTextEnhancementOptionsAdapter);
+    }
+
+    @Override
+    public void onItemClick(int position) {
+        final Enhancer enhancer = mEnhancerList.get(position);
+        enhancer.enhance();
+    }
+
+    @OnClick(R.id.createtextpdf)
+    public void openCreateTextPdf() {
+        new MaterialDialog.Builder(mActivity)
+                .title(R.string.creating_pdf)
+                .content(R.string.enter_file_name)
+                .input(getString(R.string.example), mFileNameWithType,
+                        (dialog, input) -> {
+                            if (StringUtils.getInstance().isEmpty(input)) {
+                                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_name_not_blank);
+                            } else {
+                                final String inputName = input.toString();
+                                if (!mFileUtils.isFileExist(inputName + getString(R.string.pdf_ext))) {
+                                    createPdf(inputName);
+                                } else {
+                                    MaterialDialog.Builder builder = DialogUtils.getInstance()
+                                            .createOverwriteDialog(mActivity);
+                                    builder.onPositive((dialog12, which) -> createPdf(inputName))
+                                            .onNegative((dialog1, which) -> openCreateTextPdf())
+                                            .show();
+                                }
+                            }
+                        })
+                .show();
+    }
+
+    /**
+     * function to create PDF
+     *
+     * @param mFilename name of file to be created.
+     */
+    private void createPdf(String mFilename) {
+
+        final FontSizeEnhancer fontSizeEnhancer = (FontSizeEnhancer) mEnhancerList.get(0);
+        final FontFamilyEnhancer fontFamilyEnhancer = (FontFamilyEnhancer) mEnhancerList.get(1);
+        final PasswordEnhancer passwordEnhancer = (PasswordEnhancer) mEnhancerList.get(3);
+        final FontColorEnhancer fontColorEnhancer = (FontColorEnhancer) mEnhancerList.get(4);
+        final PageColorEnhancer pageColorEnhancer = (PageColorEnhancer) mEnhancerList.get(5);
+
+        mPath = mDirectoryUtils.getOrCreatePdfDirectory().getPath();
+        mPath = mPath + "/" + mFilename + mActivity.getString(R.string.pdf_ext);
+        TextToPDFOptions options = new TextToPDFOptions(mFilename, PageSizeUtils.mPageSize,
+                passwordEnhancer.isPasswordProtected(),
+                passwordEnhancer.getPassword(),
+                mTextFileUri,
+                fontSizeEnhancer.getFontSize(),
+                fontFamilyEnhancer.getFontFamily(),
+                fontColorEnhancer.getFontColor(),
+                pageColorEnhancer.getPageColor());
+        TextToPDFUtils fileUtil = new TextToPDFUtils(mActivity);
+        new TextToPdfAsync(fileUtil, options, mFileExtension,
+                TextToPdfFragment.this).execute();
+    }
+
+    /**
+     * Create a file picker to get text file.
+     */
+    @OnClick(R.id.selectFile)
+    public void selectTextFile() {
+        if (!mPermissionGranted) {
+            getRuntimePermissions();
+            return;
+        }
+        if (mButtonClicked == 0) {
+            Uri uri = Uri.parse(Environment.getRootDirectory() + "/");
+            Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
+            intent.setDataAndType(uri, "*/*");
+            String[] mimeTypes = {"application/vnd.openxmlformats-officedocument.wordprocessingml.document",
+                    "application/msword", getString(R.string.text_type)};
+            intent.putExtra(Intent.EXTRA_MIME_TYPES, mimeTypes);
+            intent.addCategory(Intent.CATEGORY_OPENABLE);
+            try {
+                startActivityForResult(
+                        Intent.createChooser(intent, String.valueOf(R.string.select_file)),
+                        mFileSelectCode);
+            } catch (android.content.ActivityNotFoundException ex) {
+                StringUtils.getInstance().showSnackbar(mActivity, R.string.install_file_manager);
+            }
+            mButtonClicked = 1;
+        }
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        mButtonClicked = 0;
+        if (requestCode == mFileSelectCode) {
+            if (resultCode == RESULT_OK) {
+                mTextFileUri = data.getData();
+                StringUtils.getInstance().showSnackbar(mActivity, R.string.text_file_selected);
+                String fileName = mFileUtils.getFileName(mTextFileUri);
+                if (fileName != null) {
+                    if (fileName.endsWith(Constants.textExtension))
+                        mFileExtension = Constants.textExtension;
+                    else if (fileName.endsWith(Constants.docxExtension))
+                        mFileExtension = Constants.docxExtension;
+                    else if (fileName.endsWith(Constants.docExtension))
+                        mFileExtension = Constants.docExtension;
+                    else {
+                        StringUtils.getInstance().showSnackbar(mActivity, R.string.extension_not_supported);
+                        return;
+                    }
+                }
+                mFileNameWithType = mFileUtils.stripExtension(fileName) + getString(R.string.pdf_suffix);
+                mSelectFile.setText(getString(R.string.text_file_name) + fileName);
+                mCreateTextPdf.setEnabled(true);
+                mMorphButtonUtility.morphToSquare(mCreateTextPdf, mMorphButtonUtility.integer());
+            }
+        }
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        if (context instanceof Activity) {
+            mActivity = (Activity) context;
+        } else {
+            mActivity = getActivity();
+        }
+
+        mFileUtils = new FileUtils(mActivity);
+        mDirectoryUtils = new DirectoryUtils(mActivity);
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode,
+                                           @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (grantResults.length < 1)
+            return;
+        if (requestCode == PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT) {
+            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                mPermissionGranted = true;
+                selectTextFile();
+                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_permissions_given);
+            } else
+                StringUtils.getInstance().showSnackbar(mActivity, R.string.snackbar_insufficient_permissions);
+        }
+    }
+
+    private void getRuntimePermissions() {
+        PermissionsUtils.requestRuntimePermissions(
+                mActivity,
+                READ_WRITE_PERMISSIONS,
+                PERMISSION_REQUEST_WRITE_EXTERNAL_STORAGE_RESULT
+        );
+    }
+
+    @Override
+    public void onPDFCreationStarted() {
+        mMaterialDialog = DialogUtils.getInstance().createAnimationDialog(mActivity);
+        mMaterialDialog.show();
+    }
+
+    @Override
+    public void onPDFCreated(boolean success) {
+        if (mMaterialDialog != null && mMaterialDialog.isShowing())
+            mMaterialDialog.dismiss();
+        if (!success) {
+            StringUtils.getInstance().showSnackbar(mActivity, R.string.error_occurred);
+            mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
+            mCreateTextPdf.setEnabled(false);
+            mTextFileUri = null;
+            mButtonClicked = 0;
+            return;
+        }
+        StringUtils.getInstance().getSnackbarwithAction(mActivity, R.string.snackbar_pdfCreated)
+                .setAction(R.string.snackbar_viewAction, v -> mFileUtils.openFile(mPath)).show();
+        mSelectFile.setText(R.string.select_text_file);
+        mMorphButtonUtility.morphToGrey(mCreateTextPdf, mMorphButtonUtility.integer());
+        mCreateTextPdf.setEnabled(false);
+        mTextFileUri = null;
+        mButtonClicked = 0;
+    }
+
+    @Override
+    public void updateView() {
+        mTextEnhancementOptionsAdapter.notifyDataSetChanged();
+    }
+}
