+++ app/src/main/java/swati4star/createpdf/interfaces/Enhancer.java
@@ -1,0 +1,18 @@
+package swati4star.createpdf.interfaces;
+
+import swati4star.createpdf.model.EnhancementOptionsEntity;
+
+/**
+ * The {@link Enhancer} is a functional interface for all enhancements.
+ */
+public interface Enhancer {
+    /**
+     * To apply an enhancement.
+     */
+    void enhance();
+
+    /**
+     * @return The {@link EnhancementOptionsEntity} for this {@link Enhancer}.
+     */
+    EnhancementOptionsEntity getEnhancementOptionsEntity();
+}
